move unit tests for JschScpCommandProcessorImpl to common for unit tests coverage reporting
